Move FileDialogConfig to separate config.rs module
