+++ src/lib.rs
@@ -53,8 +53,11 @@
 //! }
 //! ```
 
+mod config;
 mod file_dialog;
-pub use file_dialog::{DialogMode, DialogState, FileDialog, FileDialogConfig};
-
 mod create_directory_dialog;
 mod data;
+
+pub use config::FileDialogConfig;
+pub use file_dialog::{DialogMode, DialogState, FileDialog};
+
